Refactoring Product Controller & Wiev. Dividing the controller and view Product into 2 pages.
