+++ src/main/java/ru/wms/WarehouseManagementService/controller/order/OrderController.java
@@ -1,4 +1,4 @@
-package ru.wms.WarehouseManagementService.controller;
+package ru.wms.WarehouseManagementService.controller.order;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
