+++ src/main/java/ru/wms/WarehouseManagementService/controller/product/MainProductController.java
@@ -1,12 +1,9 @@
-package ru.wms.WarehouseManagementService.controller;
+package ru.wms.WarehouseManagementService.controller.product;
 
-import jakarta.validation.Valid;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import ru.wms.WarehouseManagementService.entity.Product;
 import ru.wms.WarehouseManagementService.entity.Warehouse;
@@ -16,12 +13,11 @@ import ru.wms.WarehouseManagementService.security.UserPrincipal;
 import ru.wms.WarehouseManagementService.service.ProductService;
 import ru.wms.WarehouseManagementService.service.WarehouseService;
 
-import java.util.List;
 import java.util.Optional;
 
 @Controller
 @RequestMapping("/products")
-public class ProductController {
+public class MainProductController {
     @Autowired
     private final ProductService productService;
     @Autowired
@@ -32,7 +28,7 @@ public class ProductController {
     private WarehouseService warehouseService;
 
     @Autowired
-    public ProductController(ProductService productService, UserRepository userRepository, ProductRepository productRepository, WarehouseService warehouseService) {
+    public MainProductController(ProductService productService, UserRepository userRepository, ProductRepository productRepository, WarehouseService warehouseService) {
         this.productService = productService;
         this.userRepository = userRepository;
         this.productRepository = productRepository;
@@ -40,8 +36,7 @@ public class ProductController {
     }
 
     @GetMapping
-    public String products(Model model, @AuthenticationPrincipal UserPrincipal userPrincipal
-    ) {
+    public String products(Model model, @AuthenticationPrincipal UserPrincipal userPrincipal) {
         var user = userPrincipal.getUser();
 
         Iterable<Product> productList = productService.getAllProducts();
@@ -49,28 +44,8 @@ public class ProductController {
         model.addAttribute("warehouses", warehouseList);
         model.addAttribute("products", productList);
         model.addAttribute("product", new Product());
+
         return "products";
-    }
-
-    @PostMapping
-    public Product createProduct(
-            @ModelAttribute("product")
-            @Valid Product product,
-            @Valid Warehouse warehouse,
-            BindingResult bindingResult,
-            @AuthenticationPrincipal UserPrincipal userPrincipal
-    ) {
-        if (bindingResult.hasErrors()) {
-            throw new IllegalArgumentException("Invalid product data");
-        }
-        var user = userPrincipal.getUser();
-
-        product.setUser(user);
-        product.setWarehouse(warehouse);
-
-        Product savedProduct = productService.saveProduct(product);
-
-        return savedProduct;
     }
 
     @GetMapping("/{id}")
@@ -88,7 +63,7 @@ public class ProductController {
         return "products";
     }
 
-    @PostMapping("/{id}")
+    @PostMapping("/{id}/delete")
     public String deleteProduct(@PathVariable("id") Long id) {
         productService.deleteProductById(id);
