+++ src/main/resources/templates/createProduct.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="ru" xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+<head>
+    <meta charset="UTF-8">
+    <title>Products</title>
+</head>
+<body>
+
+<div class="container">
+
+    <h2>Вы вошли в систему как <span sec:authentication="name"></span></h2>
+
+    <hr>
+        <p>
+         Раздел управления товарами (регистрация товара)
+        </p>
+    <hr>
+
+    <form th:action="@{/createProduct}" th:object="${product}" method="post">
+        <label>Имя: <input type="text" th:field="*{name}" required /></label><br>
+        <label>Описание: <input type="text" th:field="*{description}" required /></label><br>
+        <label>Категория: <input type="text" th:field="*{category}" required /></label><br>
+        <label>Стоимость: <input type="text" th:field="*{price}" required /></label><br>
+        <label>Количество: <input type="number" th:field="*{quantity}" required /></label><br>
+        <label>Дата создания: <input type="date" th:field="*{creationDate}" required /></label><br>
+        <label>Выберите склад:
+            <select th:field="*{warehouse}" required>
+                <option th:each="warehouse : ${warehouses}" th:value="${warehouse.id}" th:text="${warehouse.name}"></option>
+            </select>
+        </label><br>
+
+        <input type="submit" value="Зарегистрировать товар" />
+        <a href="/products">Отменить регистрацию товара</a>
+    </form>
+
+    <h3><a href="/products">Вернутся в главное меню товаров</a></h3>
+
+    <h3><a href="/index">Вернутся в основное меню</a></h3>
+
+</div>
+
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+</body>
+</html>
